Delete-Admins-Some-Edits

+++ app/Http/Controllers/DeleteAccountController.php
@@ -10,32 +10,41 @@ class DeleteAccountController extends Controller
 {
     //
     private string $email = 'az540546@gmail.com';
-    public function deleteAdminAccount($id)
+    public function deleteAdminAccount(Request $request, $id)
     {
-        $admin = Admin::find($id);
-        if ($admin->email == $this->email)
-            return response()->json([
-                'icon' => 'error',
-                'title' => 'Can not delete.',
-                'text' => 'You are trying to delete supper admin account',
-            ], Response::HTTP_BAD_REQUEST);
-
-        if (!is_null($admin)) {
-            if ($admin->delete()) {
-                return response()->json([
-                    'icon' => 'success',
-                    'title' => 'Deleted',
-                    'text' => 'Account deleted successfully',
-                ], Response::HTTP_OK);
-            }else {
+        $username = explode("@", auth('admin')->user()->email);
+        if ($request->get('code') == 'delete.my.auth.account.' . $username[0] . '.sure') {
+            $admin = Admin::find($id);
+            if ($admin->email == $this->email)
                 return response()->json([
                     'icon' => 'error',
-                    'title' => 'Faild',
-                    'text' => 'Faild to delete your account',
+                    'title' => 'Can not delete.',
+                    'text' => 'You are trying to delete supper admin account',
                 ], Response::HTTP_BAD_REQUEST);
+
+            if (!is_null($admin)) {
+                if ($admin->delete()) {
+                    return response()->json([
+                        'icon' => 'success',
+                        'title' => 'Deleted',
+                        'text' => 'Account deleted successfully',
+                    ], Response::HTTP_OK);
+                }else {
+                    return response()->json([
+                        'icon' => 'error',
+                        'title' => 'Faild',
+                        'text' => 'Faild to delete your account',
+                    ], Response::HTTP_BAD_REQUEST);
+                }
+            }else {
+                return redirect()->route('admin.dashbord');
             }
         }else {
-            return redirect()->route('admin.dashbord');
+            return response()->json([
+                'icon' => 'error',
+                'title' => 'Error',
+                'text' => 'Wrong input code'
+            ], Response::HTTP_BAD_REQUEST);
         }
     }
 }
